Bug Fix + Drawing Pieces with effects
* Added gradient effects to pieces so they look nicer.

*** Bug Fix ***
Demo mode continues running in the background if you navigate to Main
Menu whilst the Demo game is in progress. This can cause weird effects
when moving into a new game mode afterwards.
The fix includes adding an ABORTED state to the GameState enum. So now,
the ReversiEngine will first check if the game has been aborted before
running it's thread.